fix: delete the "remove dialog" component and integrate the delete dialog directly into the remove component
