+++ src/app/extensions/product-notifications/shared/product-notification-edit/product-notification-edit.component.ts
@@ -16,7 +16,7 @@ import { ProductNotificationEditDialogComponent } from '../product-notification-
  *
  * @example
  * <ish-product-notification-edit
- *   [cssClass]="'btn-link btn-tool'"
+ *   cssClass="btn-link btn-tool"
  *   [productNotification]="productNotification"
  *   displayType="icon"></ish-product-notification-edit>
  */
